services/tg-gateway/requirements.txt
aiogram==3.10.0
google-generativeai==0.7.2
python-dotenv==1.0.1
aiohttp==3.9.5
